rename diagnose.py to _diagnose.py
