+++ python/taichi/_diagnose.py
