+++ python/taichi/main.py
@@ -12,7 +12,7 @@ from pathlib import Path
 
 import numpy as np
 import taichi.cc_compose
-import taichi.diagnose
+from taichi import _diagnose
 from colorama import Fore
 from taichi.core import ti_core as _ti_core
 from taichi.tools import video
@@ -714,7 +714,7 @@ class TaichiMain:
     @register
     def diagnose(arguments: list = sys.argv[2:]):
         """System diagnose information"""
-        taichi.diagnose.main()
+        _diagnose.main()
 
     @register
     def cc_compose(self, arguments: list = sys.argv[2:]):
